fixed gurobi wrapper
